+++ app/G91/gonnykh.alexander/lab2/src/main/java/ru/vsu/amm/java/classes/Truck.java
@@ -17,7 +17,7 @@ public class Truck extends Car {
     }
 
     @Override
-    public float calculatePrice() {
+    public double calculatePrice() {
         return (fourWheelDrive) ? 2 * startPrice : startPrice;
     }
